# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Eigen - BE Test Case",
        "description": "When integrating with other client, set header Accept: application/json for every request, otherwise API may not behave correctly.",
        "version": "1.0.0"
    },
    "paths": {
        "/api/v1/books": {
            "get": {
                "tags": [
                    "Books"
                ],
                "summary": "API to list all available books and its borrowing status.",
                "operationId": "books-index",
                "parameters": [
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Number of items per page",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "10"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page #",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "1"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Return books data"
                    },
                    "500": {
                        "description": "Internal server error."
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/books/{book}/borrow": {
            "post": {
                "tags": [
                    "Books"
                ],
                "summary": "API to borrow a book",
                "operationId": "book-borrow",
                "parameters": [
                    {
                        "name": "book",
                        "in": "path",
                        "description": "Book's code",
                        "required": true
                    }
                ],
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "member"
                                ],
                                "properties": {
                                    "member": {
                                        "description": "Member's code",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Book borrowed successfully"
                    },
                    "422": {
                        "description": "Validation errors."
                    },
                    "400": {
                        "description": "Bad request."
                    },
                    "500": {
                        "description": "Internal server error."
                    }
                }
            }
        },
        "/api/v1/books/{book}/return": {
            "post": {
                "tags": [
                    "Books"
                ],
                "summary": "API to return a borrowed book",
                "operationId": "book-return",
                "parameters": [
                    {
                        "name": "book",
                        "in": "path",
                        "description": "Book's code",
                        "required": true
                    }
                ],
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "member"
                                ],
                                "properties": {
                                    "member": {
                                        "description": "Member's code",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Book returned successfully"
                    },
                    "422": {
                        "description": "Validation errors."
                    },
                    "400": {
                        "description": "Bad request."
                    },
                    "500": {
                        "description": "Internal server error."
                    }
                }
            }
        },
        "/api/v1/members": {
            "get": {
                "tags": [
                    "Members"
                ],
                "summary": "API to list all members and its borrowing status.",
                "operationId": "members-index",
                "parameters": [
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Number of items per page",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "10"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page #",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "1"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Return books data"
                    },
                    "500": {
                        "description": "Internal server error."
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    }
}